Combine source files together
Add the contents of NonTiledLayer.WMS.js to NonTiledLayer.js for the following reasons:

- Reduces the final shipped file size of the library as no additonal module wrappers are created.
- Resolves an issue where both files export a function, but in the final result only NonTiledLayer is really exported since they are combined in the build.
